refactor(examples): organize optimum examples into category dirs
Remove old top-level example scripts and update PR CI to reference the
new category subdirectories (decoder_models, multimodal_models,
pooling_models).
