+++ .github/workflows/rbln_optimum_ci.yaml
@@ -125,13 +125,13 @@ jobs:
       
       - name : Run Qwen3-embedding (V1)
         run: >
-          python3 examples/optimum/run_qwen3_embedding.py
+          python3 examples/optimum/pooling_models/run_qwen3_embedding.py
           --num_input_prompt $((2 * ${{ env.NUM_INPUT_PROMPT }}))
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/qwen3-embedding-0_6b-32k_kv4k_batch1
 
       - name : Run Qwen3-reranker (V1)
         run: >
-          python3 examples/optimum/run_qwen3_reranker.py
+          python3 examples/optimum/pooling_models/run_qwen3_reranker.py
           --max_seq_len 32768
           --num_input_prompt $((2 * ${{ env.NUM_INPUT_PROMPT }}))
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/qwen3-reranker-0_6b-40k_kv4k_batch2
@@ -143,7 +143,7 @@ jobs:
 
       - name : Run text embedding model (V1)
         run: >
-          python3 examples/optimum/run_encoder_only.py
+          python3 examples/optimum/pooling_models/run_encoder_only.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/bge-m3-1k-batch4
           --q_prompt_txt ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/prompts/q_prompts.txt
@@ -227,7 +227,7 @@ jobs:
 
       - name: Run decoder-only test (eager attn) (V1)
         run: >
-          python3 examples/optimum/run_decoder_only.py
+          python3 examples/optimum/decoder_models/run_decoder_only.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/llama2-7b_batch2
           --prompt_txt ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/prompts/copy_prompts.txt
@@ -235,7 +235,7 @@ jobs:
 
       - name: Run decoder-only test (Flash-attention mode) (V1)
         run: >
-          python3 examples/optimum/run_decoder_only.py
+          python3 examples/optimum/decoder_models/run_decoder_only.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/llama3_2-3b-128k_kv16k_batch4
           --prompt_txt ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/prompts/copy_prompts.txt
@@ -243,37 +243,37 @@ jobs:
 
       - name : Run Llava-next (Eager mode) (V1)
         run: >
-           python3 examples/optimum/run_llava.py
+           python3 examples/optimum/multimodal_models/run_llava.py
            --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
            --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/llava-v1.6-mistral-7b-hf-32k-b4/
 
       - name : Run Llava-next (Flash-attention mode) (V1)
         run: >
-          python3 examples/optimum/run_llava.py
+          python3 examples/optimum/multimodal_models/run_llava.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/llava-v1.6-mistral-7b-hf-32k-b4-kv16k
 
       - name : Run Idefics3 (Eager mode) (V1)
         run: >
-          python3 examples/optimum/run_idefics3.py
+          python3 examples/optimum/multimodal_models/run_idefics3.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/idefics3-8b-llama3-32k-b4
 
       - name : Run Idefics3 (Flash-attention mode) (V1)
         run: >
-          python3 examples/optimum/run_idefics3.py
+          python3 examples/optimum/multimodal_models/run_idefics3.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/idefics3-8b-llama3-32k-b4-kv16k
 
       - name : Run Blip2 (V1)
         run: >
-          python3 examples/optimum/run_blip2.py
+          python3 examples/optimum/multimodal_models/run_blip2.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/blip2-opt-2.7b-2k-b4
 
       - name : Run PaliGemma (Eager mode) (V1)
         run: >
-          python3 examples/optimum/run_paligemma.py
+          python3 examples/optimum/multimodal_models/run_paligemma.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/paligemma-3b-8k-b4
 
@@ -354,7 +354,7 @@ jobs:
 
       - name : Run Qwen2.5_VL (V1)
         run: >
-          python3 examples/optimum/run_qwen_vl.py
+          python3 examples/optimum/multimodal_models/run_qwen_vl.py
           --num_input_prompt ${{ env.NUM_INPUT_PROMPT }}
           --model_id ${{ env.REBEL_VLLM_PRE_COMPILED_DIR }}/qwen2_5-vl-7b-32k-b4-kv16k
